Move AST selectors into own compile units
Split compare and unify also into own units.
