+++ Makefile.conf
@@ -7,6 +7,9 @@
 
 SOURCES = \
 	ast.cpp \
+	ast_sel_cmp.cpp \
+	ast_sel_unify.cpp \
+	ast_selectors.cpp \
 	node.cpp \
 	context.cpp \
 	constants.cpp \
